front-end client side form validation on submit added

+++ petManagement_BE/src/main/java/com/example/petManagement/controller/PetController.java
@@ -33,7 +33,7 @@ public class PetController {
     }
 
     @GetMapping("/find/{id}")
-    public ResponseEntity<Pet> getPetById(@PathVariable("id") Long id) {
+    public ResponseEntity<Pet> getPetById(@PathVariable("id") Integer id) {
         Pet pet = petService.findPetById(id);
         return new ResponseEntity<>(pet, HttpStatus.OK);
     }
