+++ petManagement_BE/src/main/java/com/example/petManagement/repo/PetRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PetRepo extends JpaRepository<Pet, Long>  {
-         Optional<Pet> findPetById(Long id);
+public interface PetRepo extends JpaRepository<Pet, Integer>  {
+         Optional<Pet> findPetById(Integer id);
      }
